RGB-015 octagon explosion
