+++ src/components/CreateExplosion/CreateChunk.jsx
@@ -1,9 +1,11 @@
 import BoxChunk from "./BoxChunk";
 import SphereChunk from "./SphereChunk";
+import OctagonChunk from "./OctagonChunk";
 
 const CHUNKS = {
   'sphereChunk': ({ props, key }) => <SphereChunk {...props} key={key}/>,
-  'boxChunk': ({ props, key }) => <BoxChunk {...props} key={key}/>
+  'boxChunk': ({ props, key }) => <BoxChunk {...props} key={key}/>,
+  'octagonChunk': ({ props, key }) => <OctagonChunk {...props} key={key}/>
 }
 
 const CreateChunk = ({
