+++ src/components/CreateShape/OctagonShape.jsx
@@ -1,7 +1,7 @@
-import { useRef, useEffect } from "react";
+import { useRef, useEffect, useState } from "react";
 import { useCylinder } from '@react-three/cannon';
-import { CONTACT_MATERIALS } from '../../configs/contactMaterials';
-import { COLLISION_GROUPS, OCT_QUATERNION } from '../../configs/constants';
+import { CONTACT_MATERIALS, COLLISION_GROUPS } from '../../configs/physics';
+import { OCT_QUATERNION } from '../../configs/shape';
 
 
 const OctagonShape = ({
@@ -10,6 +10,7 @@ const OctagonShape = ({
   color, 
   handleRemoveShape,
   name }) => {
+  const [visible, setVisible] = useState(false)
   const _positionRef = useRef([0, 50, 0])
   const [ref, api] = useCylinder(() => ({
     material: CONTACT_MATERIALS.SHAPE,
@@ -21,6 +22,7 @@ const OctagonShape = ({
     quaternion: OCT_QUATERNION,
     onCollide: handleOnCollide
   }), useRef(null));
+  
 
   const handleOnCollide = (e) => {
     if (e.body?.isShape && e.body.color === e.target.color)  {
@@ -28,18 +30,19 @@ const OctagonShape = ({
         name: e.target.name, 
         position: _positionRef.current,
         color: e.target.color,
-        chunkType: 'boxChunk'
+        chunkType: 'octagonChunk'
       }) 
     }
   }
 
   useEffect(() => {
+    setTimeout(() => setVisible(true), 10)
     const unsubscribe = api.position.subscribe((pos) => (_positionRef.current = pos))
     return unsubscribe
   }, [])
 
   return (
-    <mesh ref={ref} name={name} color={color} isShape={true}>
+    <mesh visible={visible} ref={ref} name={name} color={color} isShape={true}>
       <cylinderGeometry args={args}/>
       <meshStandardMaterial color={color} />
     </mesh>
