+++ src/components/CreateShape/SphereShape.jsx
@@ -1,7 +1,6 @@
 import { useRef, useEffect } from "react";
 import { useSphere } from '@react-three/cannon';
-import { CONTACT_MATERIALS } from '../../configs/contactMaterials';
-import { COLLISION_GROUPS } from '../../configs/constants';
+import { CONTACT_MATERIALS, COLLISION_GROUPS } from '../../configs/physics';
 
 const SphereShape = ({
   args,
