+++ src/components/ExplosionController/ExplosionController.jsx
@@ -1,5 +1,4 @@
 import useStore from '../../store/useStore'
-import { useState, useEffect } from 'react'
 import CreateExplosion from '../CreateExplosion/CreateExplosion'
