+++ src/components/GameStages/GameStages.jsx
@@ -1,5 +1,4 @@
 import IntroGame from './IntroGame/IntroGame'
-import SelectLevel from './SelectLevel/SelectLevel'
 import IntroLevel from './IntroLevel/IntroLevel'
 import PlayLevel from './PlayLevel/PlayLevel'
 import WinLevel from './WinLevel/WinLevel'
@@ -10,7 +9,6 @@ const GameStages = () => {
   return (
     <>
       <IntroGame/>
-      {/* <SelectLevel/> */}
       <IntroLevel/>
       <PlayLevel/>
       <WinLevel/>
