+++ src/components/GameStages/IntroGame/PuppetBox.jsx
@@ -1,7 +1,6 @@
 import { useRef } from "react";
 import { useBox } from '@react-three/cannon';
-import { COLLISION_GROUPS } from '../../../configs/constants'
-import { CONTACT_MATERIALS } from "../../../configs/contactMaterials";
+import { CONTACT_MATERIALS, COLLISION_GROUPS } from "../../../configs/physics";
 
 const PuppetBox = ({
   color,
