--- src/components/GameStages/SelectLevel/SelectLevel.jsx
@@ -1,11 +1,0 @@
-import useStore from '../../../store/useStore'
-
-
-const SelectLevel = () => {
-  const { setGameStage } = useStore((state) => state)
-
-  return null
-}
-
-
-export default SelectLevel;
