+++ src/components/GameStages/WinLevel/WinLevel.jsx
@@ -1,5 +1,5 @@
 import useStore from '../../../store/useStore'
-import { GAME_STAGES } from '../../../configs/constants'
+import { GAME_STAGES } from '../../../configs/game'
 import { useEffect } from 'react'
 
 const WinLevel = () => {
