+++ src/configs/container.js
@@ -1,0 +1,15 @@
+
+
+
+export const CONTAINER_WIDTH = 10
+
+export const CONTAINER = {
+  width: CONTAINER_WIDTH,
+  depth: CONTAINER_WIDTH * .33,
+  wallThickness: .5,
+}
+
+
+
+
+
