+++ src/configs/game.js
@@ -1,0 +1,21 @@
+export const COLORS = {
+  red: '#ff0000',
+  green: '#00ff00',
+  blue: '#3381ff'
+}
+
+export const LEVEL_ORDER = [
+  'octagons', 
+  'octBoxSphere', 
+  'boxes', 
+  'spheres', 
+  'boxSphereBox', 
+  'sphereBoxSphere'
+]
+
+export const GAME_STAGES = {
+  IntroGame: 'IntroGame',
+  IntroLevel: 'IntroLevel',
+  PlayLevel: 'PlayLevel',
+  WinLevel: 'WinLevel'
+}
